webserver/requirements.txt
aioredis==1.1.0
appnope==0.1.0
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.5.2
Automat==0.6.0
backcall==0.1.0
cachetools==2.1.0
certifi==2018.4.16
channels==2.1.1
channels-redis==2.2.1
chardet==3.0.4
constantly==15.1.0
daphne==2.1.2
decorator==4.3.0
Django==2.0.6
django-cors-headers==2.2.0
djangorestframework==3.8.2
google-auth==1.5.0
grpcio==1.12.1
grpcio-tools==1.12.1
gunicorn==19.8.1
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.22
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
kubernetes==6.0.0
msgpack==0.5.6
oauthlib==2.1.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pyasn1==0.4.3
pyasn1-modules==0.2.1
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Twisted==18.4.0
txaio==2.10.0
urllib3==1.23
wcwidth==0.1.7
websocket-client==0.48.0
zope.interface==4.5.0
